requirements.txt
# Core dependencies for Neural MMOT Solver
# Python 3.10+ required

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# JAX (for Phase 2a integration)
jax>=0.4.13
jaxlib>=0.4.13

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Data handling
pandas>=2.0.0
h5py>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Progress & Logging
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0  # Parallel testing

# Code quality
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Utilities
scikit-learn>=1.2.0  # For additional metrics
joblib>=1.2.0  # Parallel processing
